backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
passlib==1.7.4
python-dotenv==1.0.0
SQLAlchemy==2.0.4
Werkzeug==2.2.3
gunicorn==20.1.0